+++ src/app/loadout-drawer/auto-loadouts.ts
@@ -186,8 +186,7 @@ export function gatherEngramsLoadout(
  * Move a list of items to a store
  */
 export function itemMoveLoadout(items: DimItem[], store: DimStore): Loadout {
-  // Don't move things from the postmaster or that can't move
-  items = items.filter((i) => !i.location.inPostmaster && !i.notransfer);
+  items = items.filter((i) => !i.notransfer);
   items = addUpStackables(items);
 
   const itemsByType = _.mapValues(