+++ src/app/loadout/Loadouts.tsx
@@ -1,5 +1,6 @@
 import { LoadoutParameters } from '@destinyitemmanager/dim-api-types';
 import { DestinyAccount } from 'app/accounts/destiny-account';
+import { D2ManifestDefinitions } from 'app/destiny2/d2-definitions';
 import BungieImage from 'app/dim-ui/BungieImage';
 import CharacterSelect from 'app/dim-ui/CharacterSelect';
 import ClassIcon from 'app/dim-ui/ClassIcon';
@@ -7,11 +8,12 @@ import PageWithMenu from 'app/dim-ui/PageWithMenu';
 import ShowPageLoading from 'app/dim-ui/ShowPageLoading';
 import { t } from 'app/i18next-t';
 import ConnectedInventoryItem from 'app/inventory/ConnectedInventoryItem';
-import { DimItem } from 'app/inventory/item-types';
+import { DimItem, PluggableInventoryItemDefinition } from 'app/inventory/item-types';
 import ItemPopupTrigger from 'app/inventory/ItemPopupTrigger';
 import { allItemsSelector, bucketsSelector, sortedStoresSelector } from 'app/inventory/selectors';
 import { DimStore } from 'app/inventory/store-types';
 import { useLoadStores } from 'app/inventory/store/hooks';
+import { isPluggableItem } from 'app/inventory/store/sockets';
 import { getCurrentStore, getStore } from 'app/inventory/stores-helpers';
 import { SelectedArmorUpgrade } from 'app/loadout-builder/filter/ArmorUpgradePicker';
 import ExoticArmorChoice from 'app/loadout-builder/filter/ExoticArmorChoice';
@@ -19,7 +21,7 @@ import { deleteLoadout } from 'app/loadout-drawer/actions';
 import { maxLightLoadout } from 'app/loadout-drawer/auto-loadouts';
 import { applyLoadout } from 'app/loadout-drawer/loadout-apply';
 import { editLoadout } from 'app/loadout-drawer/loadout-events';
-import { Loadout } from 'app/loadout-drawer/loadout-types';
+import { DimLoadoutItem, Loadout } from 'app/loadout-drawer/loadout-types';
 import {
   convertToLoadoutItem,
   extractArmorModHashes,
@@ -47,16 +49,18 @@ import { useIsPhonePortrait } from 'app/shell/selectors';
 import { LoadoutStats } from 'app/store-stats/CharacterStats';
 import { useThunkDispatch } from 'app/store/thunk-dispatch';
 import { itemCanBeEquippedBy, itemCanBeInLoadout } from 'app/utils/item-utils';
+import { getSocketsByIndexes } from 'app/utils/socket-utils';
 import { copyString } from 'app/utils/util';
 import { DestinyClass } from 'bungie-api-ts/destiny2';
 import clsx from 'clsx';
-import { BucketHashes } from 'data/d2/generated-enums';
+import { BucketHashes, SocketCategoryHashes } from 'data/d2/generated-enums';
 import _ from 'lodash';
 import React, { useMemo, useState } from 'react';
 import { useSelector } from 'react-redux';
 import { Link } from 'react-router-dom';
 import PlugDef from './loadout-ui/PlugDef';
 import styles from './Loadouts.m.scss';
+import { createGetModRenderKey } from './mod-utils';
 
 const categoryStyles = {
   Weapons: styles.categoryWeapons,
@@ -207,16 +211,17 @@ function LoadoutRow({
   const dispatch = useThunkDispatch();
   const defs = useD2Definitions()!;
   const allItems = useSelector(allItemsSelector);
+  const getModRenderKey = createGetModRenderKey();
 
   // Turn loadout items into real DimItems, filtering out unequippable items
-  const [items, subClass, warnitems] = useMemo(() => {
+  const [items, subclass, warnitems] = useMemo(() => {
     const [items, warnitems] = getItemsFromLoadoutItems(loadout.items, defs, allItems);
     let equippableItems = items.filter((i) => itemCanBeEquippedBy(i, store, true));
-    const subClass = equippableItems.find((i) => i.bucket.hash === BucketHashes.Subclass);
-    if (subClass) {
-      equippableItems = equippableItems.filter((i) => i !== subClass);
+    const subclass = equippableItems.find((i) => i.bucket.hash === BucketHashes.Subclass);
+    if (subclass) {
+      equippableItems = equippableItems.filter((i) => i !== subclass);
     }
-    return [equippableItems, subClass, warnitems];
+    return [equippableItems, subclass, warnitems];
   }, [loadout.items, defs, allItems, store]);
 
   const savedMods = getModsFromLoadout(defs, loadout);
@@ -296,29 +301,10 @@ function LoadoutRow({
       </div>
       {loadout.notes && <div className={styles.loadoutNotes}>{loadout.notes}</div>}
       <div className={styles.contents}>
-        {(items.length > 0 || subClass || savedMods.length > 0) && (
+        {(items.length > 0 || subclass || savedMods.length > 0) && (
           <>
-            <div className={styles.subClass}>
-              {subClass ? (
-                <ItemPopupTrigger item={subClass}>
-                  {(ref, onClick) => (
-                    <ConnectedInventoryItem
-                      innerRef={ref}
-                      onClick={onClick}
-                      item={subClass}
-                      ignoreSelectedPerks
-                    />
-                  )}
-                </ItemPopupTrigger>
-              ) : (
-                <EmptyClassItem />
-              )}
-              {power !== 0 && (
-                <div className={styles.power}>
-                  <AppIcon icon={powerActionIcon} />
-                  <span>{power}</span>
-                </div>
-              )}
+            <div>
+              <Subclass defs={defs} subclass={subclass} power={power} />
             </div>
             {['Weapons', 'Armor', 'General'].map((category) => (
               <ItemCategory
@@ -331,8 +317,8 @@ function LoadoutRow({
             ))}
             {savedMods.length > 0 ? (
               <div className={styles.mods}>
-                {savedMods.map((mod, index) => (
-                  <div key={index}>
+                {savedMods.map((mod) => (
+                  <div key={getModRenderKey(mod)}>
                     <PlugDef plug={mod} />
                   </div>
                 ))}
@@ -343,6 +329,80 @@ function LoadoutRow({
           </>
         )}
       </div>
+    </div>
+  );
+}
+
+function Subclass({
+  defs,
+  subclass,
+  power,
+}: {
+  defs: D2ManifestDefinitions;
+  subclass?: DimLoadoutItem;
+  power: number;
+}) {
+  const getModRenderKey = createGetModRenderKey();
+  const plugs = useMemo(() => {
+    const plugs: PluggableInventoryItemDefinition[] = [];
+
+    if (subclass?.sockets?.categories) {
+      for (const category of subclass.sockets.categories) {
+        const showInitial =
+          category.category.hash !== SocketCategoryHashes.Aspects &&
+          category.category.hash !== SocketCategoryHashes.Fragments;
+        const sockets = getSocketsByIndexes(subclass.sockets, category.socketIndexes);
+
+        for (const socket of sockets) {
+          const override = subclass.socketOverrides?.[socket.socketIndex];
+          const initial = socket.socketDefinition.singleInitialItemHash;
+          const hash = override || (showInitial && initial);
+          const plug = hash && defs.InventoryItem.get(hash);
+          if (plug && isPluggableItem(plug)) {
+            plugs.push(plug);
+          }
+        }
+      }
+    }
+
+    return plugs;
+  }, [subclass, defs]);
+
+  return (
+    <div className={styles.subclassContainer}>
+      <div className={styles.subclass}>
+        {subclass ? (
+          <ItemPopupTrigger item={subclass}>
+            {(ref, onClick) => (
+              <ConnectedInventoryItem
+                innerRef={ref}
+                // Disable the popup when plugs are available as we are showing
+                // plugs in the loadout and they may be different to the popup
+                onClick={plugs.length ? undefined : onClick}
+                item={subclass}
+                ignoreSelectedPerks
+              />
+            )}
+          </ItemPopupTrigger>
+        ) : (
+          <EmptyClassItem />
+        )}
+        {power !== 0 && (
+          <div className={styles.power}>
+            <AppIcon icon={powerActionIcon} />
+            <span>{power}</span>
+          </div>
+        )}
+      </div>
+      {plugs.length ? (
+        <div className={styles.subclassMods}>
+          {plugs?.map((plug) => (
+            <PlugDef key={getModRenderKey(plug)} plug={plug} />
+          ))}
+        </div>
+      ) : (
+        <div className={styles.modsPlaceholder}>{t('Loadouts.Abilities')}</div>
+      )}
     </div>
   );
 }